refactoring
Took 56 minutes
